test & update getJobByID
