--- actions/getJobbyID.ts
@@ -1,19 +1,0 @@
-import prisma from "../db/prisma";
-
-const getJobById = async (formData: FormData) => {
-  const jobId = formData.get("jobID") as string;
-
-  const result = await prisma.job.findFirst({
-    select: {
-      id: false,
-    },
-    where: {
-      id: jobId,
-      isDeleted: false,
-    },
-  });
-
-  return result;
-};
-
-export default getJobById;
